AB-130: create PatientRecord folder
